+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Bugfixes
 - Fixed an issue where the import of frame states would alter the frame index
   which would de-sync the frame index with the index in the central widget.
 - Fixed an issue in the GeneralAsciiSaver where exporting data with a metadata header.
+- Fixed an issue where the splash screen would only show after the package imports/exports
+  and html documentation build.
 
 
 v25.05.16